social_media_api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
Pillow>=10.0.0   # for profile picture uploads
django-filter>=23.0